[FIX] do not set last_sale when not in customer session
The SaleOrderPaymentTransactionEventListener is sometimes called
when the transaction is confirmed by a webhook coming from the payment
acquirer. We don't want to leak the store_cache to the payment acquirer,
so we only set it if we are in a customer session.

To detect that we are in a customer session we check the sess-cart-id
HTTP header.